moved splitby to internals
