--- src/async-split-by.d.ts
@@ -1,2 +1,0 @@
-import { asyncSplitBy } from './index'
-export default asyncSplitBy
