+++ src/index.mjs
@@ -47,7 +47,6 @@ export { default as size } from './size'
 export { default as slice } from './slice'
 export { default as some } from './some'
 export { default as splitAt } from './split-at'
-export { default as splitBy } from './split-by'
 export { default as takeWhile } from './take-while'
 export { default as tap } from './tap'
 export { default as tee } from './tee'
@@ -86,7 +85,6 @@ export { default as asyncReduce } from './async-reduce'
 export { default as asyncSize } from './async-size'
 export { default as asyncSlice } from './async-slice'
 export { default as asyncSome } from './async-some'
-export { default as asyncSplitBy } from './async-split-by'
 export { default as asyncSplitAt } from './async-split-at'
 export { default as asyncTakeWhile } from './async-take-while'
 export { default as asyncTap } from './async-tap'
