--- src/split-by.d.ts
@@ -1,2 +1,0 @@
-import { splitBy } from './index'
-export default splitBy
